backend/requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.3.2  # https://github.com/python-pillow/Pillow
argon2-cffi==21.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==3.1.13  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.7.0  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.45.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==5.0.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.8.0 # https://github.com/adamchainz/django-cors-headers

drf-yasg==1.20.0
dj-rest-auth==2.1.11
djangorestframework-simplejwt==4.8.0
django_imagekit==4.0.2
django-filter==21.1
